fix(policy): enforce canonical legacy compatibility
