+++ packages/openomni/test/policy/middleware-integration.test.ts
@@ -1,36 +1,9 @@
-import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { PolicyEngine, type PolicyContext, type PolicyRegistration } from "@openomni/agent";
-import {
-  PolicyDecision,
-  type Ingress,
-  type Policy,
-  type RuntimeResource,
-} from "@openomni/protocol";
+import { beforeEach, describe, expect, test } from "bun:test";
+import type { Ingress, Policy } from "@openomni/protocol";
 import { ChannelGrantStore, Storage } from "@openomni/session";
 import { IngressAuthorityMiddleware } from "../../src/ingress/middleware/ingress-authority";
-import { ToolRuntimePolicyMiddleware } from "../../src/execution-runtime/tool/middleware/tool-runtime-policy";
-import { buildWorkerMiddleware } from "../../src/execution-runtime/middleware";
-
-const emptyUsage = { inputTokens: 0, outputTokens: 0, totalTokens: 0 };
-
-type PreDispatchContext = Omit<PolicyContext, "timing"> & {
-  readonly resourceDescriptor?: RuntimeResource.Descriptor;
-};
 
 type WorkerControlTestAction = "cancel" | "resume" | "schedule";
-type RestrictedTrustTier = "observer" | "collaborator" | "assigned_worker";
-
-function baseCtx(overrides?: Partial<PreDispatchContext>): PreDispatchContext {
-  return {
-    steps: [],
-    usage: emptyUsage,
-    turnCount: 0,
-    isCompletion: false,
-    continuationCount: 0,
-    elapsedMs: 0,
-    ...overrides,
-  };
-}
 
 function makeInboundEvent(overrides?: Partial<Ingress.InboundEvent>): Ingress.InboundEvent {
   return {
@@ -252,343 +225,5 @@ describe("IngressAuthorityMiddleware integration", () => {
 
     expect(result.event.id).toBe("evt-1");
     expect(decisions.some((decision) => decision.factsUsed?.includes("trust.manager"))).toBe(true);
-  });
-
-  test("canonical observer trust tier overrides legacy user role", async () => {
-    const decisions: Policy.PolicyDecision[] = [];
-    const event = makeInboundEvent({
-      meta: { actor: { role: "user", actorId: "act_observer", trustTier: "observer" } },
-    });
-
-    await expect(
-      IngressAuthorityMiddleware.runPreRun({
-        event,
-        coordinator: stubCoordinator,
-        onDecision: (decision) => {
-          decisions.push(decision);
-        },
-      }),
-    ).rejects.toThrow("actor is not authorized to create top-level inbound work");
-
-    expect(decisions.some((decision) => decision.factsUsed?.includes("trust.observer"))).toBe(true);
-  });
-
-  test.each([
-    "observer",
-    "collaborator",
-    "assigned_worker",
-  ] as const)("canonical %s trust tier overrides privileged legacy resident role", async (trustTier: RestrictedTrustTier) => {
-    const event = makeInboundEvent({
-      target: { kind: "worker" },
-      meta: {
-        actor: {
-          role: "resident",
-          actorId: `act_${trustTier}`,
-          trustTier,
-          trusted: true,
-          isTrustedManager: true,
-        },
-        action: "spawn",
-      },
-    });
-
-    await expect(
-      IngressAuthorityMiddleware.runPreRun({
-        event,
-        coordinator: stubCoordinator,
-      }),
-    ).rejects.toThrow("actor is not authorized to create top-level inbound work");
-  });
-
-  test("uses channel default trust tier for unregistered actors", async () => {
-    ChannelGrantStore.put({
-      id: "grant-public-observer",
-      surface: "test",
-      channel: "public",
-      kind: "trusted_channel",
-      defaultTier: "observer",
-      createdBy: "act_owner",
-    });
-    const event = makeInboundEvent({
-      channel: "public",
-      meta: { actor: { role: "user", id: "external-user-1" } },
-    });
-
-    await expect(
-      IngressAuthorityMiddleware.runPreRun({
-        event,
-        coordinator: stubCoordinator,
-      }),
-    ).rejects.toThrow("actor is not authorized to create top-level inbound work");
-  });
-
-  test("allows explicit resident target when coordinator is missing", async () => {
-    const event = makeInboundEvent({
-      target: { kind: "resident" },
-      meta: { actor: { role: "user" } },
-    });
-
-    const result = await IngressAuthorityMiddleware.runPreRun({ event });
-
-    expect(result.target.kind).toBe("resident");
-    expect(result.coordinator).toBeUndefined();
-  });
-
-  test("aborts when coordinator is missing for worker creation target", async () => {
-    const event = makeInboundEvent({ target: { kind: "worker" } });
-
-    await expect(IngressAuthorityMiddleware.runPreRun({ event })).rejects.toThrow(
-      "coordinator is required for worker target",
-    );
-  });
-
-  test("aborts on invalid schema", async () => {
-    const badEvent = { not: "valid" };
-
-    await expect(
-      IngressAuthorityMiddleware.runPreRun({
-        event: badEvent,
-        coordinator: stubCoordinator,
-      }),
-    ).rejects.toThrow();
-  });
-
-  test("aborts on unsupported mode", async () => {
-    const event = makeInboundEvent();
-    (event as Record<string, unknown>).mode = "fork";
-
-    await expect(
-      IngressAuthorityMiddleware.runPreRun({
-        event,
-        coordinator: stubCoordinator,
-      }),
-    ).rejects.toThrow();
-  });
-
-  test("collects policy decisions via onDecision callback", async () => {
-    const decisions: unknown[] = [];
-    const event = makeInboundEvent({
-      meta: { actor: { role: "user" } },
-    });
-
-    await IngressAuthorityMiddleware.runPreRun({
-      event,
-      coordinator: stubCoordinator,
-      onDecision: (d) => {
-        decisions.push(d);
-      },
-    });
-
-    expect(decisions.length).toBeGreaterThan(0);
-    for (const d of decisions) {
-      expect(d).toHaveProperty("policyId");
-      expect(d).toHaveProperty("verdict");
-    }
-  });
-
-  test("registrations produce all six middleware steps", () => {
-    const state = {
-      input: makeInboundEvent(),
-      coordinator: stubCoordinator,
-    } satisfies Parameters<typeof IngressAuthorityMiddleware.registrations>[0];
-    const regs = IngressAuthorityMiddleware.registrations(state);
-
-    expect(regs).toHaveLength(6);
-    const names = regs.map((r) => r.name);
-    expect(names).toContain(IngressAuthorityMiddleware.CoordinatorPresence.name);
-    expect(names).toContain(IngressAuthorityMiddleware.SchemaValidation.name);
-    expect(names).toContain(IngressAuthorityMiddleware.BlacklistCheck.name);
-    expect(names).toContain(IngressAuthorityMiddleware.ChannelGrantCheck.name);
-    expect(names).toContain(IngressAuthorityMiddleware.AuthorityCheck.name);
-    expect(names).toContain(IngressAuthorityMiddleware.ModeDispatch.name);
-  });
-});
-
-describe("ToolRuntimePolicyMiddleware integration", () => {
-  test("evaluates pre-tool for low-risk tier without workspace lock", async () => {
-    const result = await ToolRuntimePolicyMiddleware.evaluatePreTool({
-      toolName: "read_file",
-      toolCallId: "call-1",
-      input: { path: "/tmp/test.txt" },
-      riskTier: 0,
-      lockOwnerId: "run-1",
-    });
-
-    expect(result.decision.verdict).toBe("allow");
-    expect(result.handle.lockAcquired).toBe(false);
-    expect(result.handle.timeoutMs).toBe(30_000);
-  });
-
-  test("evaluates pre-tool for high-risk tier with timeout escalation", async () => {
-    const result = await ToolRuntimePolicyMiddleware.evaluatePreTool({
-      toolName: "bash",
-      toolCallId: "call-2",
-      input: { command: "ls" },
-      riskTier: 2,
-      lockOwnerId: "run-2",
-    });
-
-    expect(result.decision.verdict).toBe("allow");
-    expect(result.handle.timeoutMs).toBe(60_000);
-  });
-
-  test("evaluates post-tool releasing lock when acquired", async () => {
-    const handle: ToolRuntimePolicyMiddleware.RuntimePolicyHandle = {
-      timeoutMs: 30_000,
-      lockOwnerId: "run-3",
-      lockAcquired: false,
-    };
-
-    const verdict = ToolRuntimePolicyMiddleware.evaluatePostTool({
-      toolName: "read_file",
-      input: {},
-      handle,
-    });
-
-    expect(verdict.verdict).toBe("allow");
-  });
-
-  test("enforceTimeout rejects after specified ms", async () => {
-    const slow = new Promise<string>((resolve) => setTimeout(() => resolve("done"), 500));
-
-    await expect(ToolRuntimePolicyMiddleware.enforceTimeout(slow, 10)).rejects.toThrow(
-      "timeout after 10ms",
-    );
-  });
-
-  test("enforceTimeout resolves if promise completes in time", async () => {
-    const fast = Promise.resolve("ok");
-
-    const result = await ToolRuntimePolicyMiddleware.enforceTimeout(fast, 5_000);
-    expect(result).toBe("ok");
-  });
-
-  test("TimeoutError carries the timeout duration", () => {
-    const error = new ToolRuntimePolicyMiddleware.TimeoutError(42);
-    expect(error.timeoutMs).toBe(42);
-    expect(error.name).toBe("TimeoutError");
-    expect(error.message).toBe("timeout after 42ms");
-  });
-
-  test("custom timeout config overrides default tier timeouts", async () => {
-    const result = await ToolRuntimePolicyMiddleware.evaluatePreTool({
-      toolName: "write_file",
-      input: {},
-      riskTier: 1,
-      lockOwnerId: "run-4",
-      timeoutConfig: { tier1: 5_000 },
-    });
-
-    expect(result.handle.timeoutMs).toBe(5_000);
-  });
-
-  test("collects decisions via onDecision callback", async () => {
-    const decisions: unknown[] = [];
-
-    await ToolRuntimePolicyMiddleware.evaluatePreTool({
-      toolName: "bash",
-      input: { command: "echo hi" },
-      riskTier: 2,
-      lockOwnerId: "run-5",
-      onDecision: (d) => {
-        decisions.push(d);
-      },
-    });
-
-    expect(decisions.length).toBeGreaterThan(0);
-    for (const d of decisions) {
-      expect(d).toHaveProperty("verdict");
-    }
-  });
-});
-
-describe("cross-middleware deny-wins", () => {
-  beforeEach(() => {
-    Storage.reset();
-    Storage.initialize({ dbPath: ":memory:" });
-    ChannelGrantStore.put({
-      id: "grant-test",
-      surface: "test",
-      kind: "trusted_channel",
-      createdBy: "act_owner",
-    });
-  });
-
-  afterEach(() => {
-    Storage.reset();
-  });
-
-  test("deny-wins across policy engine boundaries with mixed verdicts", async () => {
-    const engine = PolicyEngine.create({ audit: false });
-
-    const allowPolicy: PolicyRegistration = {
-      name: "allow-policy",
-      timing: "invoke.prepare",
-      priority: 0,
-      failPolicy: "fail-closed",
-      fn: () => PolicyDecision.allow({ policyId: "allow-policy", reasonCodes: ["allowed"] }),
-    };
-
-    const denyPolicy: PolicyRegistration = {
-      name: "deny-policy",
-      timing: "invoke.prepare",
-      priority: 10,
-      failPolicy: "fail-closed",
-      fn: () =>
-        PolicyDecision.deny({
-          policyId: "deny-policy",
-          reasonCodes: ["denied by policy"],
-          effects: [{ type: "run.abort", reason: "denied by policy" }],
-        }),
-    };
-
-    engine.register(allowPolicy);
-    engine.register(denyPolicy);
-
-    const verdict = await engine.dispatch("invoke.prepare", {
-      ...baseCtx(),
-      toolName: "some_tool",
-      toolInput: {},
-    });
-
-    expect(verdict.verdict).toBe("deny");
-    expect(verdict.policyId).toBe("agent.policy.composed");
-    expect(verdict.reasonCodes).toContain("denied by policy");
-  });
-
-  test("worker middleware deny + tool runtime continue → overall deny", async () => {
-    const workerRegs = buildWorkerMiddleware({
-      permissions: { action: "tool.call", allowlist: ["read_file"] },
-    });
-    const engine = PolicyEngine.create({ audit: false });
-    for (const reg of workerRegs) engine.register(reg);
-
-    const verdict = await engine.dispatch("invoke.prepare", {
-      ...baseCtx(),
-      toolName: "bash",
-      toolCallId: "call-bash",
-      toolInput: { command: "rm -rf /" },
-    });
-
-    expect(verdict.verdict).toBe("deny");
-    expect(PolicyDecision.reason(verdict)).toBe("allowlist_miss");
-  });
-
-  test("ingress deny blocks entire pipeline regardless of downstream allowances", async () => {
-    const event = makeInboundEvent({
-      meta: { actor: { role: "sub_persona" } },
-    });
-
-    let ingressFailed = false;
-    try {
-      await IngressAuthorityMiddleware.runPreRun({
-        event,
-        coordinator: stubCoordinator,
-      });
-    } catch {
-      ingressFailed = true;
-    }
-
-    expect(ingressFailed).toBe(true);
   });
 });
