+++ packages/policy/src/engine/dispatch.ts
@@ -42,63 +42,78 @@ export function createPolicyEngine<TCtx extends GenericPolicyContext>(
       ...ctx,
       timing,
     });
-    const pointId =
-      compatibility.invokeCanonicalAtLegacyDispatch === undefined
+    const pointId = compatibility.resolvePointForLegacyDispatch?.(timing, fullCtx);
+    const pointSnapshot =
+      pointId === undefined
         ? undefined
-        : compatibility.resolvePointForLegacyDispatch?.(timing, fullCtx);
+        : immutablePointSnapshot<TCtx>(fullCtx, { pointId, timing });
+    const auditCtx = pointId === undefined ? fullCtx : Object.freeze({ ...fullCtx, pointId });
+    const canonicalContractFailure =
+      pointId === undefined || pointSnapshot === undefined
+        ? undefined
+        : pointSnapshot.success
+          ? validatePointContract(pointId, pointSnapshot.value)
+          : pointContractDecision(pointId, "policy.input_invalid");
     const selected =
       pointId === undefined
         ? registrations.selectLegacy(timing, fullCtx.agentType)
         : registrations.selectLegacyCompatible(timing, fullCtx.agentType, pointId);
     const decisions: Policy.PolicyDecision[] = [];
+    let canonicalContractRecorded = false;
 
     function composeAndPublish(): Policy.PolicyDecision {
       const decision = composeFinalDecision(decisions, timing, ctx.resourceDescriptor);
-      publishComposedDecision(options, timing, fullCtx, decision);
+      publishComposedDecision(options, timing, auditCtx, decision);
       return decision;
     }
 
     if (selected.length === 0) {
       const decision = PolicyDecision.allow({ policyId: COMPOSED_POLICY_ID });
-      publishComposedDecision(options, timing, fullCtx, decision);
+      publishComposedDecision(options, timing, auditCtx, decision);
       return decision;
     }
 
     for (const reg of selected) {
-      let decision: Policy.PolicyDecision;
       const startTime = Date.now();
-      try {
-        if ("kind" in reg) {
-          if (
-            pointId === undefined ||
-            compatibility.invokeCanonicalAtLegacyDispatch === undefined
-          ) {
-            continue;
-          }
-          decision = await compatibility.invokeCanonicalAtLegacyDispatch(reg, fullCtx, pointId);
+      let decision: Policy.PolicyDecision | undefined;
+      const isCanonical = "kind" in reg;
+      if (isCanonical) {
+        if (pointId === undefined || pointSnapshot === undefined) continue;
+        if (canonicalContractFailure !== undefined) {
+          if (canonicalContractRecorded) continue;
+          canonicalContractRecorded = true;
+          decision = canonicalContractFailure;
+        } else if (pointSnapshot.success) {
+          decision = await evaluateCanonical(reg, pointId, () => reg.fn(pointSnapshot.value));
         } else {
-          decision = await reg.fn(fullCtx);
+          continue;
         }
-      } catch (err) {
-        const durationMs = Date.now() - startTime;
-        const failPolicy = reg.failPolicy ?? defaultFailPolicy(timing, ctx.resourceDescriptor);
-        const error = err instanceof Error ? err : new Error(String(err));
-        publishMiddlewareError(options, timing, reg.name, error, failPolicy, durationMs);
-        if (failPolicy === "fail-open") continue;
+      } else {
+        try {
+          decision = await reg.fn(fullCtx);
+        } catch (err) {
+          const durationMs = Date.now() - startTime;
+          const failPolicy = reg.failPolicy ?? defaultFailPolicy(timing, ctx.resourceDescriptor);
+          const error = err instanceof Error ? err : new Error(String(err));
+          publishMiddlewareError(options, timing, reg.name, error, failPolicy, durationMs);
+          if (failPolicy === "fail-open") continue;
 
-        decision = middlewareErrorDecision(reg, durationMs, timing, ctx.resourceDescriptor);
+          decision = middlewareErrorDecision(reg, durationMs, timing, ctx.resourceDescriptor);
+        }
       }
 
+      if (decision === undefined) continue;
+
       const durationMs = Date.now() - startTime;
-      const normalized = normalizeDecision(
-        decision,
-        reg,
-        durationMs,
-        timing,
-        ctx.resourceDescriptor,
-      );
-      decisions.push(recordDecision(options, reg, fullCtx, normalized));
-      publishMiddlewareDebug(options, timing, reg.name, normalized.verdict, durationMs);
+      const normalized = isCanonical
+        ? decision
+        : normalizeDecision(decision, reg, durationMs, timing, ctx.resourceDescriptor);
+      const auditRegistration =
+        normalized.policyId === "policy.point.contract" ? { name: "policy.point.contract" } : reg;
+      decisions.push(recordDecision(options, auditRegistration, auditCtx, normalized));
+      if (!isCanonical) {
+        publishMiddlewareDebug(options, timing, reg.name, normalized.verdict, durationMs);
+      }
 
       if (normalized.verdict === "deny") return composeAndPublish();
     }
@@ -110,7 +125,6 @@ export function createPolicyEngine<TCtx extends GenericPolicyContext>(
     pointId: TPointId,
     ctx: DispatchPointContextGeneric<TCtx, TPointId>,
   ): Promise<Policy.PolicyDecision> {
-    const contract = Policy.PolicyPoint.Registry[pointId];
     const timing = timingForPolicyPoint(pointId);
     const snapshot = immutablePointSnapshot(ctx, { pointId, timing });
     const auditCtx = snapshot.success ? snapshot.value : Object.freeze({ pointId, timing });
@@ -135,15 +149,7 @@ export function createPolicyEngine<TCtx extends GenericPolicyContext>(
     }
     const fullCtx = snapshot.value;
 
-    const missingContext = contract.requiredContext.some(
-      (key) => Reflect.get(fullCtx, key) === undefined,
-    );
-    let contractFailure: Policy.PolicyDecision | undefined;
-    if (missingContext) {
-      contractFailure = pointContractDecision(pointId, "policy.context_missing");
-    } else if (!Policy.PolicyPoint.InputSchemas[pointId].safeParse(fullCtx).success) {
-      contractFailure = pointContractDecision(pointId, "policy.input_invalid");
-    }
+    const contractFailure = validatePointContract(pointId, fullCtx);
     if (contractFailure !== undefined) {
       decisions.push(
         recordDecision(options, { name: "policy.point.contract" }, fullCtx, contractFailure),
@@ -161,40 +167,52 @@ export function createPolicyEngine<TCtx extends GenericPolicyContext>(
     }
 
     for (const reg of selected) {
-      const startTime = Date.now();
-      let middlewareDecision: unknown;
-      let engineDecision: Policy.PolicyDecision | undefined;
-      try {
-        middlewareDecision = await reg.fn(fullCtx);
-      } catch (err) {
-        const durationMs = Date.now() - startTime;
-        const failPolicy = reg.failPolicy ?? contract.defaultFailPolicy;
-        const error = err instanceof Error ? err : new Error(String(err));
-        publishMiddlewareError(options, timing, reg.name, error, failPolicy, durationMs);
-        if (failPolicy === "fail-open") continue;
-        engineDecision = pointMiddlewareErrorDecision(reg, pointId, durationMs);
-      }
-
-      const durationMs = Date.now() - startTime;
-      const normalized =
-        engineDecision === undefined
-          ? normalizePointDecision(middlewareDecision, reg, pointId, durationMs)
-          : { decision: engineDecision };
-      const declaredEffects = declaredEffectsFor(reg, pointId, contract.allowedEffects);
-      const undeclared = normalized.parsed?.effects.find(
-        (effect) => !declaredEffects.some((declared) => declared === effect.type),
-      );
-      const enforced =
-        undeclared === undefined
-          ? normalized.decision
-          : undeclaredEffectDecision(reg, pointId, undeclared.type, durationMs);
+      const enforced = await evaluateCanonical(reg, pointId, () => reg.fn(fullCtx));
+      if (enforced === undefined) continue;
       decisions.push(recordDecision(options, reg, fullCtx, enforced));
-      publishMiddlewareDebug(options, timing, reg.name, enforced.verdict, durationMs);
 
       if (enforced.verdict === "deny") return composeAndPublish();
     }
 
     return composeAndPublish();
+  }
+
+  async function evaluateCanonical(
+    reg: PolicyEngineRegistrationGeneric<TCtx>,
+    pointId: PolicyPointId,
+    invoke: () => Promise<Policy.PolicyDecision> | Policy.PolicyDecision,
+  ): Promise<Policy.PolicyDecision | undefined> {
+    const contract = Policy.PolicyPoint.Registry[pointId];
+    const timing = timingForPolicyPoint(pointId);
+    const startTime = Date.now();
+    let middlewareDecision: unknown;
+    let engineDecision: Policy.PolicyDecision | undefined;
+    try {
+      middlewareDecision = await invoke();
+    } catch (err) {
+      const durationMs = Date.now() - startTime;
+      const failPolicy = reg.failPolicy ?? contract.defaultFailPolicy;
+      const error = err instanceof Error ? err : new Error(String(err));
+      publishMiddlewareError(options, timing, reg.name, error, failPolicy, durationMs);
+      if (failPolicy === "fail-open") return undefined;
+      engineDecision = pointMiddlewareErrorDecision(reg, pointId, durationMs);
+    }
+
+    const durationMs = Date.now() - startTime;
+    const normalized =
+      engineDecision === undefined
+        ? normalizePointDecision(middlewareDecision, reg, pointId, durationMs)
+        : { decision: engineDecision };
+    const declaredEffects = declaredEffectsFor(reg, pointId, contract.allowedEffects);
+    const undeclared = normalized.parsed?.effects.find(
+      (effect) => !declaredEffects.some((declared) => declared === effect.type),
+    );
+    const enforced =
+      undeclared === undefined
+        ? normalized.decision
+        : undeclaredEffectDecision(reg, pointId, undeclared.type, durationMs);
+    publishMiddlewareDebug(options, timing, reg.name, enforced.verdict, durationMs);
+    return enforced;
   }
 
   return {
@@ -204,6 +222,20 @@ export function createPolicyEngine<TCtx extends GenericPolicyContext>(
     dispatch,
     dispatchPoint,
   };
+}
+
+function validatePointContract(
+  pointId: PolicyPointId,
+  fullCtx: Readonly<object>,
+): Policy.PolicyDecision | undefined {
+  const contract = Policy.PolicyPoint.Registry[pointId];
+  if (contract.requiredContext.some((key) => Reflect.get(fullCtx, key) === undefined)) {
+    return pointContractDecision(pointId, "policy.context_missing");
+  }
+  if (!Policy.PolicyPoint.InputSchemas[pointId].safeParse(fullCtx).success) {
+    return pointContractDecision(pointId, "policy.input_invalid");
+  }
+  return undefined;
 }
 
 function declaredEffectsFor<TCtx extends GenericPolicyContext>(
